fixed float cluster lables when using mask

+++ src/spike_sort/core/cluster.py
@@ -102,7 +102,7 @@ def cluster(method, features,  *args, **kwargs):
     if mask is not None:
         valid_data = data[mask,:]
         cl = cluster_func(valid_data, *args, **kwargs)
-        labels = np.zeros(data.shape[0])-1
+        labels = np.zeros(data.shape[0], dtype='int')-1
         labels[mask] = cl
     else:
         labels = cluster_func(data, *args, **kwargs)
